+++ interfaces/HO-Portal/src/app/shared/password-toggle-input/password-toggle-input.component.ts
@@ -12,7 +12,7 @@ import {
   styleUrls: ['./password-toggle-input.component.scss'],
 })
 export class PasswordToggleInputComponent {
-  @ViewChild('passwordInput')
+  @ViewChild('passwordInput', { static: true })
   public passwordInput: any;
 
   @Input()